lineapp/requirements.txt

Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1

psycopg2-binary==2.9.9

cryptography==41.0.7
PyJWT[crypto]==2.8.0  
requests==2.31.0

python-jose==3.3.0
ulid-py==1.1.0
Pillow==10.4.0
colorama==0.4.6